fixed existing tests
